+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - `SharedObject`:
   - Renamed `isIsolateCopy` to `isAuxiliaryInstance`.
+- Renamed `SharedType` to `ReferenceableType`.
+- Renamed `SharedStoreIsolateServer` to `SharedStoreIsolateMain`.
+- Renamed `SharedStoreIsolateClient` to `SharedStoreIsolateAuxiliary`.
+- Renamed `SharedMapIsolateServer` to `SharedMapIsolateMain`.
+- Renamed `SharedMapIsolateClient` to `SharedMapIsolateAuxiliary`.
 
 ## 1.0.10